:fire: Divide response funcs into each file
