+++ lib/dialogue/misc.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+from lib import misc
+
+RESPONSES = (
+    "ああ(;´Д`)",
+    "ﾏｼﾞで？(;´Д`)",
+    "ﾜﾗﾀ(;´Д`)",
+    "何言ってるんだこの人(;´Д`)",
+    "ﾔﾊﾞｲな(;´Д`)",
+    "畏れ(;´Д`)",
+    "わかってるよ(;´Д`)",
+    "つーかまんこだろ"
+)
+
+def _random_choice(*arg):
+    return misc.choice(RESPONSES)
